+++ CHANGELOG.md
@@ -5,6 +5,7 @@ NOTE: isort follows the [semver](https://semver.org/) versioning standard.
 
 ### 5.1.4 July 19, 2020
   - Fixed issue #1333: Use of wrap_length raises an exception about it not being lower or equal to line_length.
+  - Fixed issue #1330: Ensure stdout can be stubbed dynamically for `show_unified_diff` function.
 
 ### 5.1.3 July 18, 2020
   - Fixed issue #1329: Fix comments duplicated when --fass option is set.